+++ src/ImGui/Control/Window/WindowFlags.cs
@@ -1,0 +1,47 @@
+﻿using System;
+
+namespace ImGui
+{
+    [Flags]
+    public enum WindowFlags
+    {
+        Default = 0,
+        NoTitleBar = 1 << 0,   // Disable title-bar
+        NoResize = 1 << 1,   // Disable user resizing with the lower-right grip
+        NoMove = 1 << 2,   // Disable user moving the window
+        NoScrollbar = 1 << 3,   // Disable scrollbars (window can still scroll with mouse or programatically)
+        NoScrollWithMouse = 1 << 4,   // Disable user vertically scrolling with mouse wheel
+        NoCollapse = 1 << 5,   // Disable user collapsing window by double-clicking on it
+        AlwaysAutoResize = 1 << 6,   // Resize every window to its content every frame
+        ShowBorders = 1 << 7,   // Show borders around windows and items
+        NoSavedSettings = 1 << 8,   // Never load/save settings in .ini file
+        NoInputs = 1 << 9,   // Disable catching mouse or keyboard inputs
+        MenuBar = 1 << 10,  // Has a menu-bar
+        HorizontalScrollbar = 1 << 11,  // Allow horizontal scrollbar to appear (off by default). You may use SetNextWindowContentSize(ImVec2(width,0.0f)); prior to calling Begin() to specify width. Read code in imgui_demo in the "Horizontal Scrolling" section.
+        NoFocusOnAppearing = 1 << 12,  // Disable taking focus when transitioning from hidden to visible state
+        NoBringToFrontOnFocus = 1 << 13,  // Disable bringing window to front when taking focus (e.g. clicking on it or programatically giving it focus)
+        AlwaysVerticalScrollbar = 1 << 14,  // Always show vertical scrollbar (even if ContentSize.y < Size.y)
+        AlwaysHorizontalScrollbar = 1 << 15,  // Always show horizontal scrollbar (even if ContentSize.x < Size.x)
+        AlwaysUseWindowPadding = 1 << 16,  // Ensure child windows without border uses style.WindowPadding (ignored by default for non-bordered child windows, because more convenient)
+
+        VerticalScrollbar = 1 << 17,
+
+        // [Internal]
+        ChildWindow = 1 << 20,  // Don't use! For internal use by BeginChild()
+        ChildWindowAutoFitX = 1 << 21,  // Don't use! For internal use by BeginChild()
+        ChildWindowAutoFitY = 1 << 22,  // Don't use! For internal use by BeginChild()
+        ComboBox = 1 << 23,  // Don't use! For internal use by ComboBox()
+        Tooltip = 1 << 24,  // Don't use! For internal use by BeginTooltip()
+        Popup = 1 << 25,  // Don't use! For internal use by BeginPopup()
+        Modal = 1 << 26,  // Don't use! For internal use by BeginPopupModal()
+        ChildMenu = 1 << 27   // Don't use! For internal use by BeginMenu()
+    };
+
+    internal static class WindowFlagsExtension
+    {
+        public static bool HaveFlag(this WindowFlags value, WindowFlags flag)
+        {
+            return (value & flag) != 0;
+        }
+    }
+}
